requirements.txt
alembic==1.6.5
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
executing==0.8.0
fastapi==0.68.0
google-api-core==2.1.0
google-auth==1.34.0
google-cloud-appengine-logging==0.2.0
google-cloud-audit-log==0.1.1
google-cloud-core==2.1.0
google-cloud-logging==2.6.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.41.0
h11==0.12.0
httptools==0.1.2
icecream==2.1.1
idna==2.10
importlib-metadata==4.6.1
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
mysqlclient==2.0.3
packaging==21.0
pluggy==0.13.1
proto-plus==1.19.0
protobuf==3.17.2
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
socketlabs-injectionapi==1.2.1
SQLAlchemy==1.4.17
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
zipp==3.5.0
